Set mean productivity.

Set mean productivity (Mg/ha/yr) across all census intervals for a given site.

determine_mean_prod(DF, site, flag_stranglers, exclude_interval)

Arguments

DF
A data.table.

site
Provide the full name of your site (in lower case) i.e. 'barro colorado island'.

exclude_interval
A vector (i.e. c(1,2)) indicating if a set of census intervals must be discarded from computation due for instance to a change in protocol of measurment.

strangler
TRUE or FALSE, if TRUE strangler figs tree are flagged (upon a list to published soon).

Value

Mean productivity in Mg/ha/yr.